split components into separate files
